requirements/base.txt
# Bleeding edge Django
django==1.8.6

# Configuration
django-environ==0.4.0
django-secure==1.0.1

# Models
django-model-utils==2.3.1

# Images
Pillow==3.0.0

# Python-PostgreSQL Database Adapter
psycopg2==2.6.1

# Time zones support
pytz==2015.7

# Redis support
django-redis==4.3.0
redis>=2.10.0

celery==3.1.19

# Your custom requirements go here

django-appconf==1.0.1
djangocms-admin-style==1.0.6
django-admin-shortcuts==1.2.6
djangocms-youtube==0.2.0
django-wysiwyg-redactor==0.4.9
jsonfield==1.0.3
embedly==0.5.0


# django-filer (1.0.3) - PyPi version has migration issues
-e git+https://github.com/stefanfoulis/django-filer.git@master#egg=django-filer

# django-cms (3.2.0) — PyPi version has migration issues
-e git+https://github.com/divio/django-cms.git@master#egg=django-cms

